+++ example/lib/config.dart
@@ -28,7 +28,7 @@ class Config {
   static const Environment environment = Environment.test;
   static const String baseUrl = "checkout-test.adyen.com";
   static const String apiVersion = "v71";
-  static const String iOSReturnUrl = "flutter-ui-host://payments";
+  static const String iOSReturnUrl = "adyencheckout://com.adyen.adyen_checkout_example";
   static const GooglePayEnvironment googlePayEnvironment =
       GooglePayEnvironment.test;