+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-## Next version
+## 1.2.0
 
 * Added sessionResult to the PaymentSessionFinished payment result.
+* Added the loading bottom sheet for the advanced flow google pay component.
+* Improved dynamic viewport of card component.
+* Updated iOS SDK to v5.14.0.
+* Updated Android SDK to v5.8.0. Gradle v8 is now mandatory.
 
 ## 1.1.0